Lezárt bérlésekben hozzáadásra került a napidijak figyelembevételével a számlázás fő motorja. Ez visszadja a bérlés alatti parkolási perceket, vezetési perceket. A teljes bérlési időintervallum figyelembevétele mellett parkolási-vezetési-parkolási időt külön nézi és külön is kalkulálja. Meghatározásra került, hogy amennyiben a vezetés+parkolás+indítás+kényelmi díj esetlegesen magasabb lenne, mint a napi bérlésnek az összege, úgy a napibérlés legyen a végösszeg. Arazas táblába felkerült minden előfizetési csomaghoz mérten a saját díjai. Ugyanígy az előfizetési fajták az Elofizetesek táblában. Továbbá a NapiBerlesek táblába felkerültek a 2.naptól egészen a 20. napig minden autófajta ára, minden előfizetéses használatra vetítve. További fejlesztésre szorul még a hibakeresési mechanika, továbbá a számlázási felület kialakítása, illetve amint ez elkészül a lezart_berlesek táblából átintegrálásra kerülnek a funkciók a számlázási táblába.

+++ fullstack_0.2/backend/database/migrations/2024_11_17_180100_create_kategoriak_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('kategoriak', function (Blueprint $table) {
             $table->id('kat_id');
-            $table->integer('kat_besorolas')->unique(); 
+            $table->unsignedBigInteger('kat_besorolas')->unique();
             $table->timestamps();
         });
     }
